requirements.txt
fastapi
uvicorn
sqlalchemy
duckdb
duckdb-engine
rapidfuzz
openai
pypdf
python-docx
python-multipart
